Create DrawingBook.py
In this question it is said that there is a book, 1st page of the book is always on right side, and as you turn page, page no. 2 and 3 will be there, and if last page is even it is on left side, and if last page is odd it is on right side, you have to tell how many min. page turns required to reach to certain page, you can start turning pages from the front or from the back.